backend/requirements.txt
# Flask Framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Database
pyodbc==5.0.1

# Security  
bcrypt==4.1.2
PyJWT==2.8.0

# Validation
marshmallow==3.20.1

# Utilities
python-dateutil==2.8.2